Factory pattern abuses

+++ base/VkSynObjectBundleBase.cpp
@@ -1,6 +1,6 @@
 #include "VkSynObjectBundleBase.h"
 
-VkSynObjectBundleBase::VkSynObjectBundleBase(uint8_t _bundle_size) :
+VkSynObjectBundleBase::VkSynObjectBundleBase(uint32_t _bundle_size) :
     bundle_size(_bundle_size)
 {
 }
